Turn MailboxSession into POJO & delete SimpleMailboxSession
